refactor: login screen
